# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transport Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: hubId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"

    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{staffId}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{staffId}/toggle-active:
    patch:
      operationId: toggleStaffActive
      tags: [staff]
      summary: Toggle staff active status
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"

  # PARCELS
...
components:
  schemas:
...
    StaffMember:
      type: object
      required: [id, name, phone, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        hubId:
          type: ["integer", "null"]
        hubName:
          type: ["string", "null"]
        hubCode:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, phone, email, password, role]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        hubId:
          type: ["integer", "null"]
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        hubId:
          type: ["integer", "null"]
...