# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transport Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffInput {
  name: string;
  phone: string;
  email: string;
  password: string;
  role: string;
  /** @nullable */
  hubId?: number | null;
}